requirements.txt
streamlit
PyMuPDF
langchain
langchain-community
langchain-core
langchain-ollama
python-docx
reportlab
